add some error handling for literal parsing
